Add files via upload
My copy of it, didn't want to overwrite what you guys had just in case, but only minor changes were made to implement the bullet pool with it. Tested as much as I could without the VR equipment and no errors happened, but shouldn't be any different cause only difference should be the method of input, all the logic is the same. By the way, in gun script add the bullet object being set active to true, that's only change I made in that script.